backend/upload-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pillow==10.1.0
aiofiles==23.2.0
python-multipart==0.0.6
minio==7.2.0
pika==1.3.2
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2